+++ js-miniapp-bridge/test/test.spec.ts
@@ -28,6 +28,8 @@ import {
   SecureStorageUnavailableError,
   SecureStorageIOError,
   EsimConfig,
+  PermissionName,
+  PermissionStatus,
 } from '../src';
 
 /* tslint:disable:no-any */
@@ -1140,6 +1142,122 @@ describe('setupAndInstallEsim', () => {
   });
 });
 
+describe('forceLogout', () => {
+  it('will receive result of requesting force logout', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, true);
+
+    return expect(
+      bridge.userProfileManager.forceLogout()
+    ).to.eventually.deep.equal(true);
+  });
+
+  it('will parse error', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, '{ "message": "test message" }');
+
+    return expect(bridge.userProfileManager.forceLogout()).to.eventually.be
+      .rejected;
+  });
+});
+
+describe('forceInternalWebView', () => {
+  it('will receive result of requesting internal web view', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, true);
+
+    return expect(
+      bridge.webviewConfigManager.forceInternalWebView(true)
+    ).to.eventually.deep.equal(true);
+  });
+
+  it('will receive result of requesting internal web view 2', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, true);
+
+    return expect(bridge.forceInternalWebView(true)).to.eventually.deep.equal(
+      true
+    );
+  });
+
+  it('will parse error', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, '{ "message": "test message" }');
+
+    return expect(bridge.webviewConfigManager.forceInternalWebView(true)).to
+      .eventually.be.rejected;
+  });
+});
+
+describe('requestPermissionStatus', () => {
+  it('will call the platform executor', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+
+    bridge.utilityManager
+      .getPermissionStatus(PermissionName.MICROPHONE)
+      .catch(handleError);
+
+    sinon.assert.calledWith(mockExecutor.exec, 'getPermissionStatus');
+  });
+
+  it('will parse the Permission JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, `granted`);
+
+    return expect(
+      bridge.utilityManager.getPermissionStatus(PermissionName.CAMERA)
+    ).to.eventually.deep.equal(PermissionStatus.GRANTED);
+  });
+
+  it('will parse the Error response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(3, '{ "message": "Gallery not found" }');
+
+    return expect(
+      bridge.utilityManager.getPermissionStatus(PermissionName.GALLERY)
+    ).to.eventually.be.rejected;
+  });
+});
+
+describe('launchAppSettings', () => {
+  it('will call the platform executor', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+
+    bridge.utilityManager.launchAppSettings().catch(handleError);
+
+    sinon.assert.calledWith(mockExecutor.exec, 'launchAppSettings');
+  });
+
+  it('will parse the enabled JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, `true`);
+
+    return expect(
+      bridge.utilityManager.launchAppSettings()
+    ).to.eventually.deep.equal(true);
+  });
+
+  it('will parse the disabled JSON response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(2, `false`);
+
+    return expect(
+      bridge.utilityManager.launchAppSettings()
+    ).to.eventually.deep.equal(false);
+  });
+
+  it('will parse the Error response', () => {
+    const bridge = new Bridge.MiniAppBridge(mockExecutor);
+    mockExecutor.exec.callsArgWith(
+      3,
+      '{"message":"Launch App Settings not found"}'
+    );
+
+    return expect(bridge.utilityManager.launchAppSettings()).to.eventually.be
+      .rejected;
+  });
+});
+
 interface CreateCallbackParams {
   onSuccess?: (success: any) => any;
   onError?: (error: string) => any;
@@ -1152,6 +1270,6 @@ function createCallback({
   return {
     onSuccess: onSuccess || (() => undefined),
     onError: onError || (() => undefined),
-    id: String(Math.random()),
+    id: String(Bridge.cryptoRandom()),
   };
 }